+++ backend/canisters/user/impl/src/lib.rs
@@ -208,12 +208,9 @@ struct Data {
 }
 
 fn event_sink_client() -> EventSinkClient<CdkRuntime> {
-    EventSinkClientBuilder::new(
-        CanisterId::from_text("nq4qv-wqaaa-aaaaf-bhdgq-cai").unwrap(),
-        CdkRuntime::default(),
-    )
-    .with_flush_delay(Duration::from_millis(5 * MINUTE_IN_MS))
-    .build()
+    EventSinkClientBuilder::new(ic_cdk::caller(), CdkRuntime::default())
+        .with_flush_delay(Duration::from_millis(5 * MINUTE_IN_MS))
+        .build()
 }
 
 impl Data {